Move new accessor to private headers
